% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spec.R
\name{plot_spec}
\alias{plot_spec}
\title{Create a line chart presenting specific country/countries COVID information
within a time period}
\usage{
plot_spec(
  df,
  location = c("Canada"),
  val = "new_cases",
  date_from = NULL,
  date_to = NULL,
  title = NULL
)
}
\arguments{
\item{df}{Data frame of the selected covid data from get_data()}

\item{location}{Charactor vector of target country names. By default c('Canada)}

\item{val}{Quantitative values to be aggregated. Must be numeric variable.
Also known as a 'measure'. By default 'new_cases'}

\item{date_from}{Start date of the data range with format like '2021-10-31'.
By default it represents 7 days prior to today's date}

\item{date_to}{End date of data range with format like '2021-10-31'. By default it represents today's date}

\item{title}{The title of the plot. By default will be generated based on val}
}
\value{
A ggplot line chart
}
\description{
Create a line chart presenting specific country/countries COVID information
within a time period
}
\examples{
plot_spec(df, date_from = "2022-01-01", date_to = "2022-01-14")
}
