% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Title}
\usage{
get_data(date_from, date_to, location)
}
\arguments{
\item{date_from}{Start date of the data range with format '\%Y-\%m-\%d'. By default it represents 7 days prior to today's date}

\item{date_to}{End date of data range with format '\%Y-\%m-\%d'. By default it represents today's date}

\item{location}{Charactor vector of target country names. By default it retrieves all countries}
}
\value{
data.frame
}
\description{
Title
}
\examples{
get_data(date_from='2022-01-01', date_to="2022-01-07", location=c('Canada'))
}
