% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metric.R
\name{plot_metric}
\alias{plot_metric}
\title{Plot a line chart using ggplot}
\usage{
plot_metric(metric = "positive_rate", date_from = NULL, date_to = NULL)
}
\arguments{
\item{metric}{A character vector with, at most, one element.}

\item{date_from}{A character vector having the starting date, one element.}

\item{date_to}{A character vector having the end date, one element.}
}
\value{
A ggplot line chart
}
\description{
Plot a line chart using ggplot
}
\examples{
x <- "positive_rate"
plot_metric(x, date_from = "2022-01-01", "2022-01-15")

y <- "death_rate"
plot_metric(y, date_from = "2022-01-01", "2022-01-15")
}
